+++ plonky2_sha256/src/main.rs
@@ -21,7 +21,7 @@ pub fn prove_sha256(msg: &[u8]) -> Result<()> {
     type C = PoseidonGoldilocksConfig;
     type F = <C as GenericConfig<D>>::F;
     let mut builder = CircuitBuilder::<F, D>::new(CircuitConfig::standard_recursion_config());
-    let targets = sha256_circuit(&mut builder, len as u64);
+    let targets = sha256_circuit(&mut builder, len);
     let mut pw = PartialWitness::new();
     let expected_res = array_to_bits(hash.as_slice());
 
@@ -32,7 +32,6 @@ pub fn prove_sha256(msg: &[u8]) -> Result<()> {
     for i in 0..expected_res.len() {
         pw.set_bool_target(targets.digest[i], expected_res[i]);
 
-//TODO! 
         match expected_res[i] {
             true => builder.assert_one(targets.digest[i].target),
             false => builder.assert_zero(targets.digest[i].target),
